clk: Make lpc18xx-creg a regular kernel driver
The lpc18xx-creg clock driver requires a syscon regmap to access the registers.
However, the syscon is probed at postcore_initcall, after clk_of_init() is
called (clocks and timers need to be up early).

The lpc18xx-creg can deferred its probing, but to do so it needs to be a
regular kernel driver. This commit converts it.

Signed-off-by: Ezequiel Garcia <ezequiel@vanguardiasur.com.ar>