clk: lpc18xx-ccu: Allow to ignore unused CCU USB0 clock
For some reason, the LPC4337 cpu halts when clk_is_enabled() is called
for USB0. This happens during the unused clocks disabling initcall.

Let's add a workaround to make USB0 not disabled, but it seems a bit
unusual that the halt happens when clk_is_enabled() and not in the
disabled itself.

Notice that I was powering the board from USB.

Signed-off-by: Ezequiel Garcia <ezequiel@vanguardiasur.com.ar>